requirements.txt
numpy
tqdm
argparse
python-dotenv